+++ examples/codes/xprotocol_with_goplugin_example/make_codec.sh
@@ -1,11 +1,18 @@
 #!/bin/bash
 
-go build --buildmode=plugin  -o codec.so  \
-  ./codec/api.go  \
-  ./codec/command.go  \
-  ./codec/decoder.go  \
-  ./codec/encoder.go  \
-  ./codec/mapping.go  \
-  ./codec/matcher.go  \
-  ./codec/protocol.go \
-  ./codec/types.go
+function make_mosn {
+	mkdir ./build_mosn
+	cp ../../../cmd/mosn/main/* ./build_mosn
+	cd ./build_mosn
+	go build -o mosn
+	mv mosn ../
+	cd ../
+	rm -rf ./build_mosn
+}
+
+function make_so {
+	go build --buildmode=plugin  -o codec.so ./codec.go
+}
+
+make_so
+make_mosn
